+++ lib/philiprehberger/retry_kit.rb
@@ -21,5 +21,6 @@ end
 
 require_relative "retry_kit/version"
 require_relative "retry_kit/backoff"
+require_relative "retry_kit/budget"
 require_relative "retry_kit/circuit_breaker"
 require_relative "retry_kit/executor"